Revert "Merge branch 'master' of https://github.com/entdark/jaMME"
This reverts commit e286c4fcd9fc3f9422374fd648ac0882516cbe72, reversing
changes made to 32b4ec4d747fe3be32132ba81cf19bf93c585be2.

+++ codemp/qcommon/cm_load.cpp
@@ -1140,7 +1140,7 @@ int CM_ModelContents_Actual( clipHandle_t model, clipMap_t *cm )
 {
 	cmodel_t	*cmod;
 	int			contents = 0;
-	int64_t		i;
+	int			i;
 
 	if (!cm)
 	{
